dev-tools/dev-browser/requirements.txt
PyQt6
PyQt6-WebEngine
cryptography
markdown

# Puppeteer CDP bridge (Python WebSocket client used by puppeteer_bridge.py)
websocket-client

# macOS WKWebView (native WebKit engine)
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-WebKit
pyobjc-framework-Quartz
pyobjc-framework-AppKit

# Build the .app bundle (optional -- only needed when running build_app.sh)
# pyinstaller
